rename roles.py to messageManager.py
